+++ measure/network.py
@@ -1,6 +1,6 @@
 from flask import Flask, Blueprint, jsonify
 from flask_restplus import Api, Resource, fields
-import psutil
+import nw_service as service
 
 api_v1_network = Blueprint('api_v1_network', __name__, url_prefix='/api/1/network')
 
@@ -10,53 +10,34 @@ api = Api(api_v1_network, version='1.0', title='Network REST API',
 
 ns = api.namespace('nw', description='Network REST API')
 
-def getNIC():
-    return {"net_io_counters" : psutil.net_io_counters()._asdict()}
 
 @ns.route('/ioCounters')
-class getNetworkIoCounters(Resource):
+class NetworkIoCounters(Resource):
 
     def get(self):
-        return jsonify(getNIC())
+        return jsonify(service.getNIC())
 
-def getIS():
-    ans = {}
-    data = psutil.net_if_stats()
-    for st in data:
-        ans[st] = data[st]._asdict()
-    return {"net_if_stats" : ans}
 @ns.route('/ifStats')
-class getNetworkIfStats(Resource):
+class NetworkIfStats(Resource):
 
     def get(self):
 
-        return jsonify(getIS())
-
-def getNIA():
-    ans = {}
-    data = psutil.net_if_addrs()
-    for key in data.keys():
-        arr = []
-        for i in range(0, len(data[key])):
-
-            arr.append(data[key][i]._asdict())
-        ans[key] = arr
-    return {"net_if_addrs" : ans}
+        return jsonify(service.getIS())
 
 @ns.route('/ifAddrs')
-class getNetworkIfAddrs(Resource):
+class NetworkIfAddrs(Resource):
 
     def get(self):
-        return jsonify(getNIA())
+        return jsonify(service.getNIA())
 
 @ns.route('/netInfo')
-class getNetworkIfAddrs(Resource):
+class NetInfo(Resource):
 
     def get(self):
         ans = []
-        ans.append(getNIC())
-        ans.append(getNIA())
-        ans.append(getIS())
+        ans.append(service.getNIC())
+        ans.append(service.getNIA())
+        ans.append(service.getIS())
         return jsonify(net_info=ans)
